refactor(analysis): decompose index.go into 4 files (<500 LOC each)
Split the 732-LOC index.go into:
- index.go (222 LOC): core types, symbolIndex interface, indexer struct,
  lifecycle (Refresh, loadPackages, updateState), Lookup, Packages,
  IsSymbolUsed
- index_harvest.go (293 LOC): AST harvester for symbol/usage collection
- index_search.go (150 LOC): FindImplementors, SearchSymbols, GetUsages
- index_impl.go (105 LOC): implementation computation

Pure structural change — no semantic modifications. All tests pass
with -race, go vet clean, verify_architecture shows zero new edges,
find_usages confirms zero caller changes.
