+++ internal/tools/analysis/index.go
@@ -1,20 +1,15 @@
+// Copyright (c) 2026 gosharplite@gmail.com
+// SPDX-License-Identifier: MIT
+
 package analysis
 
 import (
 	"context"
-	"fmt"
-	"go/ast"
 	"go/token"
-	"go/types"
-	"os"
 	"path/filepath"
-	"runtime"
-	"strings"
 	"sync"
 	"time"
 
-	"golang.org/x/sync/errgroup"
-	"golang.org/x/sync/semaphore"
 	"golang.org/x/tools/go/packages"
 )
 
@@ -140,133 +135,6 @@ func (idx *indexer) Refresh(ctx context.Context, hb chan<- struct{}) error {
 	return nil
 }
 
-// pkgResult holds the results of harvesting symbols and usages from a package.
-type pkgResult struct {
-	symbols map[string][]symbolLocation
-	usages  map[string][]location
-}
-
-// sendResult sends a pkgResult on the results channel, emitting a heartbeat on
-// success. It respects context cancellation.
-func sendResult(ctx context.Context, results chan<- pkgResult, res pkgResult, hb chan<- struct{}) error {
-	select {
-	case results <- res:
-		if hb != nil {
-			select {
-			case hb <- struct{}{}:
-			default:
-			}
-		}
-		return nil
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
-func (idx *indexer) harvestPackages(ctx context.Context, fset *token.FileSet, pkgs []*packages.Package, hb chan<- struct{}) (map[string][]symbolLocation, map[string][]location, error) {
-	results := make(chan pkgResult, len(pkgs))
-	g, gCtx := errgroup.WithContext(ctx)
-
-	limit := int64(runtime.NumCPU())
-	if limit < 1 {
-		limit = 1
-	}
-	sem := semaphore.NewWeighted(limit)
-
-	for _, pkg := range pkgs {
-		p := pkg // Captured for closure
-		g.Go(func() error {
-			if err := sem.Acquire(gCtx, 1); err != nil {
-				return err
-			}
-			defer sem.Release(1)
-
-			res, err := idx.processPackage(gCtx, fset, p)
-			if err != nil {
-				return err
-			}
-			return sendResult(gCtx, results, res, hb)
-		})
-	}
-
-	// Wait for all workers to finish and close the results channel
-	go func() {
-		_ = g.Wait()
-		close(results)
-	}()
-
-	if err := g.Wait(); err != nil {
-		return nil, nil, err
-	}
-
-	return idx.mergeResults(results)
-}
-
-func (idx *indexer) processPackage(ctx context.Context, fset *token.FileSet, pkg *packages.Package) (pkgResult, error) {
-	h := newHarvester(fset)
-	h.info = pkg.TypesInfo
-
-	for _, file := range pkg.Syntax {
-		if ctx.Err() != nil {
-			return pkgResult{}, ctx.Err()
-		}
-		if err := idx.processFile(fset, file, h); err != nil {
-			return pkgResult{}, err
-		}
-	}
-
-	return pkgResult{
-		symbols: h.symbolsByPath,
-		usages:  h.usagesByName,
-	}, nil
-}
-
-func (idx *indexer) processFile(fset *token.FileSet, file *ast.File, h *harvester) error {
-	filename := fset.File(file.Pos()).Name()
-	absPath, err := filepath.Abs(filename)
-	if err != nil {
-		return err
-	}
-
-	// Although the file is already parsed into memory, we may want to skip it based on metadata
-	// or perform validation.
-	info, err := os.Stat(absPath)
-	if err != nil {
-		// In some environments, files might be deleted between Load and Refresh
-		return nil
-	}
-
-	if !idx.shouldIndexFile(absPath, info) {
-		return nil
-	}
-
-	h.currentPath = absPath
-	ast.Inspect(file, h.visit)
-	return nil
-}
-
-func (idx *indexer) shouldIndexFile(path string, info os.FileInfo) bool {
-	if info.IsDir() {
-		return false
-	}
-	return strings.HasSuffix(path, ".go")
-}
-
-func (idx *indexer) mergeResults(results <-chan pkgResult) (map[string][]symbolLocation, map[string][]location, error) {
-	symbolsByPath := make(map[string][]symbolLocation)
-	usagesByName := make(map[string][]location)
-
-	for res := range results {
-		for path, symbols := range res.symbols {
-			symbolsByPath[path] = symbols
-		}
-		for name, locations := range res.usages {
-			usagesByName[name] = append(usagesByName[name], locations...)
-		}
-	}
-	return symbolsByPath, usagesByName, nil
-}
-
 func (idx *indexer) toLocation(pos token.Pos) location {
 	p := idx.fset.Position(pos)
 	abs, _ := filepath.Abs(p.Filename)
@@ -294,143 +162,6 @@ func (idx *indexer) Lookup(ctx context.Context, symbol string, hb chan<- struct{
 	return locations, nil
 }
 
-func (idx *indexer) FindImplementors(ctx context.Context, interfaceName string, hb chan<- struct{}) ([]typeName, error) {
-	if err := idx.Refresh(ctx, hb); err != nil {
-		return nil, err
-	}
-	idx.mu.RLock()
-	defer idx.mu.RUnlock()
-
-	iface := idx.findInterfaceType(interfaceName)
-	if iface == nil {
-		return nil, fmt.Errorf("interface %s not found", interfaceName)
-	}
-
-	return idx.collectImplementors(iface), nil
-}
-
-func (idx *indexer) findInterfaceType(interfaceName string) *types.Interface {
-	for _, pkg := range idx.pkgs {
-		obj := pkg.Types.Scope().Lookup(interfaceName)
-		if obj == nil {
-			continue
-		}
-		if t, ok := obj.Type().Underlying().(*types.Interface); ok {
-			return t
-		}
-	}
-	return nil
-}
-
-func (idx *indexer) collectImplementors(iface *types.Interface) []typeName {
-	var implementors []typeName
-	for _, pkg := range idx.pkgs {
-		scope := pkg.Types.Scope()
-		for _, name := range scope.Names() {
-			obj := scope.Lookup(name)
-			if obj == nil {
-				continue
-			}
-			// Skip other interfaces
-			if _, ok := obj.Type().Underlying().(*types.Interface); ok {
-				continue
-			}
-
-			if types.Implements(obj.Type(), iface) || types.Implements(types.NewPointer(obj.Type()), iface) {
-				implementors = append(implementors, typeName{
-					PkgPath: pkg.PkgPath,
-					Name:    name,
-				})
-			}
-		}
-	}
-	return implementors
-}
-
-func (idx *indexer) SearchSymbols(ctx context.Context, path string, query string, exportedOnly bool, hb chan<- struct{}) ([]symbolLocation, error) {
-	if err := idx.Refresh(ctx, hb); err != nil {
-		return nil, err
-	}
-	idx.mu.RLock()
-	defer idx.mu.RUnlock()
-
-	searchPath, err := filepath.Abs(path)
-	if err != nil {
-		searchPath = path
-	}
-
-	var results []symbolLocation
-	query = strings.ToLower(query)
-
-	for p, syms := range idx.symbolsByPath {
-		if !idx.isInSearchPath(searchPath, p) {
-			continue
-		}
-
-		for _, sym := range syms {
-			if idx.matchesQuery(sym, query, exportedOnly) {
-				results = append(results, sym)
-			}
-		}
-	}
-	return results, nil
-}
-
-func (idx *indexer) isInSearchPath(targetPath, filePath string) bool {
-	if targetPath == filePath {
-		return true
-	}
-	rel, err := filepath.Rel(targetPath, filePath)
-	if err != nil {
-		return false
-	}
-	return !strings.HasPrefix(rel, "..")
-}
-
-func (idx *indexer) matchesQuery(sym symbolLocation, query string, exportedOnly bool) bool {
-	if exportedOnly && !ast.IsExported(sym.Name) {
-		return false
-	}
-	if query != "" && !strings.Contains(strings.ToLower(sym.Name), query) {
-		return false
-	}
-	return true
-}
-
-func (idx *indexer) GetUsages(ctx context.Context, symbol string, path string, hb chan<- struct{}) ([]location, error) {
-	if err := idx.Refresh(ctx, hb); err != nil {
-		return nil, err
-	}
-	idx.mu.RLock()
-	defer idx.mu.RUnlock()
-
-	searchPath, err := filepath.Abs(path)
-	if err != nil {
-		searchPath = path
-	}
-
-	var results []location
-	usages, ok := idx.usagesByName[symbol]
-	if !ok {
-		return nil, nil
-	}
-
-	for _, loc := range usages {
-		// Optimized path check instead of filepath.Rel
-		if loc.Path == searchPath {
-			results = append(results, loc)
-			continue
-		}
-		if strings.HasPrefix(loc.Path, searchPath) {
-			if len(loc.Path) > len(searchPath) && loc.Path[len(searchPath)] == filepath.Separator {
-				results = append(results, loc)
-			}
-		}
-	}
-
-	return results, nil
-}
-
 func (idx *indexer) Packages(ctx context.Context, hb chan<- struct{}) ([]*packages.Package, error) {
 	if err := idx.Refresh(ctx, hb); err != nil {
 		return nil, err
@@ -438,153 +169,6 @@ func (idx *indexer) Packages(ctx context.Context, hb chan<- struct{}) ([]*packag
 	idx.mu.RLock()
 	defer idx.mu.RUnlock()
 	return idx.pkgs, nil
-}
-
-type harvester struct {
-	fset          *token.FileSet
-	symbolsByPath map[string][]symbolLocation
-	usagesByName  map[string][]location
-	currentPath   string
-	info          *types.Info
-}
-
-func newHarvester(fset *token.FileSet) *harvester {
-	return &harvester{
-		fset:          fset,
-		symbolsByPath: make(map[string][]symbolLocation),
-		usagesByName:  make(map[string][]location),
-	}
-}
-
-func (h *harvester) visit(n ast.Node) bool {
-	if n == nil {
-		return true
-	}
-
-	switch d := n.(type) {
-	case *ast.GenDecl:
-		h.handleGenDecl(d)
-	case *ast.FuncDecl:
-		h.handleFuncDecl(d)
-	case *ast.Ident:
-		h.handleIdent(d)
-	}
-	return true
-}
-
-func (h *harvester) handleGenDecl(d *ast.GenDecl) {
-	for _, spec := range d.Specs {
-		switch s := spec.(type) {
-		case *ast.ValueSpec:
-			h.handleValueSpec(d, s)
-		case *ast.TypeSpec:
-			h.handleTypeSpec(s)
-		}
-	}
-}
-
-func (h *harvester) handleValueSpec(d *ast.GenDecl, s *ast.ValueSpec) {
-	kind := "var"
-	if d.Tok == token.CONST {
-		kind = "const"
-	}
-	for _, name := range s.Names {
-		loc := h.toLocation(name.Pos())
-		h.symbolsByPath[h.currentPath] = append(h.symbolsByPath[h.currentPath], symbolLocation{
-			location: loc,
-			Name:     name.Name,
-			Kind:     kind,
-		})
-	}
-}
-
-func (h *harvester) handleTypeSpec(s *ast.TypeSpec) {
-	loc := h.toLocation(s.Name.Pos())
-	h.symbolsByPath[h.currentPath] = append(h.symbolsByPath[h.currentPath], symbolLocation{
-		location: loc,
-		Name:     s.Name.Name,
-		Kind:     "type",
-	})
-}
-
-func (h *harvester) handleFuncDecl(d *ast.FuncDecl) {
-	kind := "func"
-	sig := getFuncSignature(d)
-	var recv string
-	if d.Recv != nil && len(d.Recv.List) > 0 {
-		recv = exprToString(d.Recv.List[0].Type)
-	}
-	loc := h.toLocation(d.Name.Pos())
-	h.symbolsByPath[h.currentPath] = append(h.symbolsByPath[h.currentPath], symbolLocation{
-		location:  loc,
-		Name:      d.Name.Name,
-		Kind:      kind,
-		Signature: sig,
-		Receiver:  recv,
-	})
-}
-
-func (h *harvester) handleIdent(d *ast.Ident) {
-	if h.info == nil {
-		return
-	}
-
-	if _, isDef := h.info.Defs[d]; isDef {
-		return
-	}
-
-	obj, ok := h.info.Uses[d]
-	if !ok || obj == nil {
-		return
-	}
-
-	isExported, isMethod, isPackageLevel := h.classifyObject(obj)
-	if !isExported && !isMethod && !isPackageLevel {
-		return
-	}
-
-	// Move toLocation after the filter to avoid unnecessary allocations
-	loc := h.toLocation(d.Pos())
-	h.recordUsage(d.Name, obj, loc, isExported)
-}
-
-func (h *harvester) classifyObject(obj types.Object) (isExported, isMethod, isPackageLevel bool) {
-	isExported = obj.Exported()
-
-	if fn, ok := obj.(*types.Func); ok {
-		sig := fn.Type().(*types.Signature)
-		isMethod = sig.Recv() != nil
-	}
-
-	if obj.Pkg() != nil && obj.Parent() == obj.Pkg().Scope() {
-		isPackageLevel = true
-	}
-
-	return
-}
-
-func (h *harvester) recordUsage(shortName string, obj types.Object, loc location, isExported bool) {
-	if isExported {
-		// Only call expensive getSymbolIdentity for exported symbols
-		key := getSymbolIdentity(obj)
-		h.usagesByName[key] = append(h.usagesByName[key], loc)
-		if key != shortName {
-			h.usagesByName[shortName] = append(h.usagesByName[shortName], loc)
-		}
-	} else {
-		// Private method call or package-level symbol: store by short name to save memory
-		h.usagesByName[shortName] = append(h.usagesByName[shortName], loc)
-	}
-}
-
-func (h *harvester) toLocation(pos token.Pos) location {
-	p := h.fset.Position(pos)
-	// Optimization: use h.currentPath which is already absolute
-	return location{
-		Path:   h.currentPath,
-		Line:   p.Line,
-		Column: p.Column,
-	}
 }
 
 func (idx *indexer) needsRefresh() bool {
@@ -625,100 +209,6 @@ func (idx *indexer) updateState(pkgs []*packages.Package, symbolsByPath map[stri
 	idx.usagesByName = usagesByName
 	idx.implementations = impls
 	idx.lastRefresh = time.Now()
-}
-
-func (idx *indexer) collectInterfaces(pkgs []*packages.Package) []*types.Interface {
-	var interfaces []*types.Interface
-	for _, pkg := range pkgs {
-		scope := pkg.Types.Scope()
-		for _, name := range scope.Names() {
-			obj := scope.Lookup(name)
-			if tn, ok := obj.(*types.TypeName); ok {
-				if itf, ok := tn.Type().Underlying().(*types.Interface); ok {
-					interfaces = append(interfaces, itf)
-				}
-			}
-		}
-	}
-	return interfaces
-}
-
-func (idx *indexer) asConcreteNamedType(obj types.Object) (*types.Named, bool) {
-	tn, ok := obj.(*types.TypeName)
-	if !ok {
-		return nil, false
-	}
-	named, ok := tn.Type().(*types.Named)
-	if !ok {
-		return nil, false
-	}
-	if _, ok := named.Underlying().(*types.Interface); ok {
-		return nil, false
-	}
-	return named, true
-}
-
-func (idx *indexer) mapTypeToInterfaces(impls map[string][]string, named *types.Named, interfaces []*types.Interface, pkgTypes *types.Package) {
-	for _, itf := range interfaces {
-		implements := types.Implements(named, itf) || types.Implements(types.NewPointer(named), itf)
-
-		if !implements {
-			implements = idx.satisfiesGenericInterface(named, itf, pkgTypes)
-		}
-
-		if implements {
-			idx.recordInterfaceImplementation(impls, named, itf, pkgTypes)
-		}
-	}
-}
-
-func (idx *indexer) satisfiesGenericInterface(named *types.Named, itf *types.Interface, pkgTypes *types.Package) bool {
-	if itf.NumMethods() == 0 {
-		return false
-	}
-	for i := 0; i < itf.NumMethods(); i++ {
-		m := itf.Method(i)
-		if obj, _, _ := types.LookupFieldOrMethod(named, true, pkgTypes, m.Name()); obj == nil {
-			return false
-		}
-	}
-	return true
-}
-
-func (idx *indexer) recordInterfaceImplementation(impls map[string][]string, named *types.Named, itf *types.Interface, pkgTypes *types.Package) {
-	for i := 0; i < itf.NumMethods(); i++ {
-		im := itf.Method(i)
-		imId := getSymbolIdentity(im)
-
-		cm, _, _ := types.LookupFieldOrMethod(named, true, pkgTypes, im.Name())
-		if cm != nil {
-			cmId := getSymbolIdentity(cm)
-			impls[imId] = append(impls[imId], cmId)
-		}
-	}
-}
-
-func (idx *indexer) computeImplementations(pkgs []*packages.Package) map[string][]string {
-	impls := make(map[string][]string)
-	ifaces := idx.collectInterfaces(pkgs)
-
-	for _, pkg := range pkgs {
-		for _, obj := range pkg.TypesInfo.Defs {
-			if named, ok := idx.asConcreteNamedType(obj); ok {
-				idx.mapTypeToInterfaces(impls, named, ifaces, pkg.Types)
-			}
-		}
-	}
-	return impls
-}
-
-func (idx *indexer) GetImplementations(ctx context.Context, interfaceMethodId string, hb chan<- struct{}) []string {
-	if err := idx.Refresh(ctx, hb); err != nil {
-		return nil
-	}
-	idx.mu.RLock()
-	defer idx.mu.RUnlock()
-	return idx.implementations[interfaceMethodId]
 }
 
 func (idx *indexer) IsSymbolUsed(ctx context.Context, name string, hb chan<- struct{}) bool {
